Fixing small bugs and preparing for merge.
- Added Classes module
- Updated reqs and gitignore
- Updated logging
- Added creating dumps for first run
