+++ .gitignore
@@ -4,4 +4,3 @@ logs
 venv
 .env
 *.db
-chats.txt
