+++ requirements.txt
@@ -1,7 +1,8 @@
 aiogram==2.19
+aioschedule==0.5.2
 beautifulsoup4==4.10.0
+coloredlogs==15.0.1
 pydantic==1.9.0
+python-dotenv==0.19.2
 requests==2.27.1
 lxml==4.8.0
-python-dotenv==0.19.2
-coloredlogs==15.0.1
